+++ api/server.go
@@ -354,6 +354,8 @@ func (g *GRPCServer) Start(ctx context.Context, lis net.Listener) error {
 		governanceService:         g.governanceService,
 		transfersService:          g.transferService,
 		delegationService:         g.delegationService,
+		marketService:             g.marketsService,
+		marketDepthService:        g.marketDepthService,
 	}
 
 	protoapi2.RegisterTradingDataServiceServer(g.srv, tradingDataSvcV2)